+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # adrez-mapper Revisions
 
 2020-11-26
-- add: location.isPrimary 
+- add: location.isPrimary: true, false, master, useMaster, leave (case insensitive) 
 
 2020-11-25
 - add: contact.sortName where McAnser, Jay is translated to Jay McAnser